Use double wherever float was previously used

+++ src/hw/beeper.c
@@ -18,8 +18,8 @@ static const i16 SOUND_FREQUENCIES[4] =
 
 static void sound_push_back(f8_beeper_t *beeper, unsigned frequency)
 {
-  unsigned current_tick = PF_SAMPLES * ((float)beeper->current_cycles /
-                                        (float)beeper->total_cycles);
+  unsigned current_tick = PF_SAMPLES * ((double)beeper->current_cycles /
+                                        (double)beeper->total_cycles);
 
   if (current_tick != beeper->last_tick)
   {
@@ -27,9 +27,9 @@ static void sound_push_back(f8_beeper_t *beeper, unsigned frequency)
 
     for (i = beeper->last_tick; i < current_tick; i++)
       beeper->frequencies[i] = beeper->frequency_last;
- }
- beeper->last_tick = current_tick;
- beeper->frequency_last = SOUND_FREQUENCIES[frequency];
+  }
+  beeper->last_tick = current_tick;
+  beeper->frequency_last = SOUND_FREQUENCIES[frequency];
 }
 
 F8D_OP_OUT(beeper_out)
@@ -70,8 +70,8 @@ void beeper_finish_frame(f8_device_t *device)
     else
     {
       /* Use sine wave to tell if our square wave is on or off */
-      float sine = pf_wave((2 * PF_PI * m_beeper->frequencies[i] *
-                            m_beeper->time * PF_PERIOD), FALSE);
+      double sine = pf_wave((2 * PF_PI * m_beeper->frequencies[i] *
+                             m_beeper->time * PF_PERIOD), FALSE);
       int mult = sine > 0 ? 1 : 0;
 
       m_beeper->samples[2 * i] = m_beeper->amplitude * mult;
